requirements.txt
aws-cdk-lib==2.24.1
constructs>=10.0.0,<11.0.0
pyjwkest==1.4.2
boto3>=1.23.6
